# packages/backend/openapi/openapi.yaml
...
openapi: '3.1.0'
info:
  title: RONL Business API
  # NO `version` here, deliberately. scripts/build-openapi.cjs sets it from
  # package.json and REFUSES TO BUILD if this file sets one, so the published
  # document and the release can never disagree.
  summary: Process, decision and document services for Regels.overheid.nl
  contact:
    name: Steven Gort
    email: steven.gort@ictu.nl
    url: https://iou-architectuur.open-regels.nl/
  description: |
    Business API behind Regels.overheid.nl: Operaton process and decision
    execution, task handling, document generation and signing, and the
    machine-to-machine surface consumed outside the browser applications.

    ## Authentication

    Most operations require a Keycloak-issued JWT as a bearer token, applied by
    `jwtMiddleware` inside each router rather than at the mount. A tenant is
    derived from the token and enforced per operation; some operations
    additionally require a role.

    The unauthenticated surface is deliberate and small: `/v1/health`,
    `/v1/openapi.json`, `/v1/public`, `/v1/media-aggregator`, and the ValidSign
    callback, which carries no token because ValidSign does not send one.

    ## Error responses

    Every 4xx and 5xx is RFC 9457 problem details, served as
    `application/problem+json` (#216): `type`, `status`, `title`, `detail` and
    `instance`, plus a `code` extension member -- the stable identifier to
    branch on (see the `Problem` schema). Until #216 errors answered
...

...

...

...
servers:
...
paths:
...
  /mcp/sources:
    get:
      operationId: listMcpSources
      tags: [MCP]
      summary: The MCP sources this service can reach
      responses:
        '200':
          description: The sources and whether each is connected.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PlainEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/McpSource' }
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The caller has neither `caseworker` nor `admin` (`FORBIDDEN`).
          $ref: '#/components/responses/Error'
  /mcp/models:
    get:
      operationId: listMcpModels
      tags: [MCP]
      summary: The language models available for chat
      responses:
        '200':
          description: The models.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PlainEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/McpModel' }
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The caller has neither `caseworker` nor `admin` (`FORBIDDEN`).
          $ref: '#/components/responses/Error'
  /mcp/chat:
    post:
      operationId: mcpChat
      tags: [MCP]
      summary: Ask a question across the connected MCP sources
      description: |
        **The only streaming operation in this API.** A successful call answers
        `text/event-stream` with `Cache-Control: no-cache` and
        `X-Accel-Buffering: no`, then writes one SSE `data:` line per event:

        ```
        data: {"type":"delta","text":"ok"}

        data: {"type":"done"}
        ```

        A `delta` carries the next fragment of the answer; `done` ends the
        stream. OpenAPI cannot describe an event sequence, so the 200 below
        types the body as a string and the framing is given here.

        The REFUSALS are ordinary JSON, not events — they are sent before the
        stream begins, so a caller must check the status and content type
        before reading.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message, modelId]
              properties:
                message: { type: string, description: The question. Must not be blank. }
                modelId:
                  type: string
                  description: 'One of the ids from `GET /mcp/models`.'
                history:
                  type: array
                  description: Earlier turns, oldest first.
                  items:
                    type: object
                    required: [role, content]
                    properties:
                      role: { type: string, enum: [user, assistant] }
                      content: { type: string }
                sources:
                  type: array
                  description: |
                    Restrict to these source ids. Absent or empty means every
                    connected source.
                  items: { type: string }
      responses:
        '200':
          description: The answer, streamed as server-sent events.
          content:
            text/event-stream:
              schema:
                type: string
                description: 'A sequence of `data:` lines; see the description for the event
                  shapes.'
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
        '400':
          description: |
            `message` is blank or `modelId` is absent (`INVALID_REQUEST`). The
            message names which.
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The caller has neither `caseworker` nor `admin` (`FORBIDDEN`).
          $ref: '#/components/responses/Error'
        '503':
          description: |
            MCP is switched off (`MCP_DISABLED`), or none of the selected
            sources is connected (`MCP_NOT_CONNECTED`).
          $ref: '#/components/responses/Error'
components:
  headers:
    ApiVersion:
      description: |
        The full released version of this API, set on every response by
        version.middleware.ts. NL API Design Rules, API-57.
      schema:
        type: string
        examples: ['2026.09.11']
  securitySchemes:
    m2mOAuth:
      type: oauth2
      description: |
        The machine-to-machine surface, `/v1/m2m`.

        The token is obtained with a CLIENT CREDENTIALS grant against the same
        realm the rest of the API uses, and is validated the same way: same
        issuer, `ronl-business-api` audience, RS256. It is a separate scheme
        here because the CALLER is checked as well as the token.

        **The client must be allow-listed.** `requireM2mClient` requires the
        token's `azp` -- the party it was issued to -- to appear on the server's
        `M2M_ALLOWED_CLIENTS`, which is `operaton-mcp-client` alone today.
        Anything else is refused with `403 M2M_CLIENT_NOT_ALLOWED` before any
        handler runs. Adding a consumer means adding its client id to that
        variable; nothing changes in Keycloak.

        `azp` rather than the audience, because every person in the realm holds
        a token for `ronl-business-api`: the audience cannot distinguish a
        machine from a citizen, and before #237 that meant any authenticated
        user reached all 18 operations.

        OpenAPI cannot express the allow-list condition, so it is stated here
        and carried by the 403 on every operation. A bearer token that is
        merely valid is NOT sufficient.

        `tokenUrl` below is production's. Acceptance issues tokens at
        `https://acc.keycloak.open-regels.nl/realms/ronl/protocol/openid-connect/token`.
      flows:
        clientCredentials:
          tokenUrl: https://keycloak.open-regels.nl/realms/ronl/protocol/openid-connect/token
          scopes: {}

    mediaAggregatorKey:
      type: http
      scheme: bearer
      description: |
        A shared secret for the media aggregator, NOT a Keycloak token. Checked
        only when the server sets MEDIA_AGGREGATOR_ACCEPT_KEY; when it does not,
        /media-aggregator/search is open. Applies to that operation alone.
...
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 9457 problem details, served as `application/problem+json` (#216).
        Every 4xx and 5xx this API answers has this shape.

        `code` is an extension member: the stable, machine-readable identifier
        a consumer should branch on, the same value the pre-#216 envelope
        carried as `error.code`. `title` is derived from it, so the same code
        always has the same title. An operation may document further extension
        members (`reserved`, `tenants`, ...).
      required: [type, status, title, detail, instance]
      properties:
        type:
          type: string
          description: Always `about:blank`; this API has no page per problem kind.
          examples: [about:blank]
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: The HTTP status, repeated.
        title:
          type: string
          description: A short summary of the kind of problem, derived from `code`.
        detail:
          type: string
          description: What went wrong with this request.
        instance:
          type: string
          description: The request path, without its query string.
        code:
          type: string
          description: A stable, machine-readable code, e.g. `TENANT_MISMATCH`.
      additionalProperties: true
...
    PlainEnvelope:
      type: object
      description: |
        The plainest envelope in this API: success and a payload, with no
        `meta` and no `timestamp`. Used by the execution core and by the board
        views here.
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: {}
...
    McpSource:
      type: object
      required: [id, displayName, connected]
      properties:
        id: { type: string }
        displayName: { type: string }
        description: { type: [string, 'null'] }
        connected:
          type: boolean
          description: |
            False means chat cannot use it; selecting only disconnected
            sources answers 503.
    McpModel:
      type: object
      required: [id, displayName, providerId]
      properties:
        id: { type: string, description: 'Pass to `POST /mcp/chat` as `modelId`.' }
        displayName: { type: string }
        providerId: { type: string }
        providerDisplayName: { type: [string, 'null'] }
...
  responses:
...
    Unauthorized:
      description: |
        No usable bearer token. Two codes, both from `jwtMiddleware`:
        **`MISSING_TOKEN`** when the `Authorization` header is absent or is not
        a bearer, and **`INVALID_TOKEN`** when a token is present but does not
        verify against the realm's JWKS -- expired, wrong issuer or audience,
        bad signature.

        Said `UNAUTHORIZED` until #214. That code is real, but it comes from
        the role and assurance guards further in, never from the token check
        this response describes.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Error:
      description: The request failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        API-Version:
          $ref: '#/components/headers/ApiVersion'

# Applied to every operation that does not override it with `security: []`.
# The unauthenticated surface is the exception and says so per operation.
security:
  - bearerAuth: []